requirements.txt
cv==1.0.0
Flask==1.1.2
h5py==2.10.0
imageio==2.8.0
importlib-metadata==1.6.1
Jinja2==2.11.2
Keras==2.4.2
Keras-Preprocessing==1.1.2
numpy==1.18.5
opencv-python==4.2.0.34
pandas==1.0.5
Pillow==7.1.2
pylint==2.5.3
scikit-learn==0.23.1
scipy==1.4.1
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
